Refactored Tests to use Swift Testing
